requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-flex-fields==1.0.0
importlib-metadata==5.0.0
Markdown==3.4.1
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.6
simplejwt==2.0.1
sqlparse==0.4.3
typing==3.7.4.3
tzdata==2022.6
zipp==3.10.0
